+++ features/uploader/controller.js
@@ -1,5 +1,6 @@
 const UploaderService = require('./service');
 const Busboy = require('busboy');
+const ApiError = require('../../lib/ApiError');
 const { ResponseMessage, formatUser, formatUpload, parseProjection } = require('../utils');
 
 const Logger = require('log-my-ass');
@@ -7,6 +8,10 @@ const log = new Logger(API_CONFIG.logger, 'Uploader');
 
 class UploadController {
   async add(req, res, next) {
+    if (req.ability.cannot('write', 'Upload')) {
+      return next(ApiError.Forbidden('Вы не можете загружать файлы!'));
+    }
+
     const busboy = new Busboy({ headers: req.headers });
 
     let fileUploadId, extension;
@@ -38,6 +43,10 @@ class UploadController {
   }
 
   async delete(req, res, next) {
+    if (req.ability.cannot('write', 'Upload')) {
+      return next(ApiError.Forbidden('Вы не можете удалять файлы!'));
+    }
+
     try {
       await UploaderService.delete(req.user._id, req.body.uploadId);
